Split AppConstants
