+++ core/src/main/java/nl/nn/adapterframework/management/bus/endpoints/BrowseJdbcTable.java
@@ -54,7 +54,7 @@ import nl.nn.adapterframework.util.XmlUtils;
 public class BrowseJdbcTable extends BusEndpointBase {
 
 	private static final String DB2XML_XSLT = "xml/xsl/BrowseJdbcTableExecute.xsl";
-	private static final String JDBC_PERMISSION_RULES = AppConstants.getInstance().getResolvedProperty("browseJdbcTable.permission.rules");
+	private static final String JDBC_PERMISSION_RULES = AppConstants.getInstance().getProperty("browseJdbcTable.permission.rules");
 	private static final String COLUMN_NAME = "COLUMN_NAME";
 	private static final String DATA_TYPE = "DATA_TYPE";
 	private static final String COLUMN_SIZE = "COLUMN_SIZE";
