+++ core/src/main/java/nl/nn/adapterframework/util/flow/FlowDiagramManager.java
@@ -56,8 +56,8 @@ public class FlowDiagramManager implements ApplicationContextAware, Initializing
 	private static Logger log = LogUtil.getLogger(FlowDiagramManager.class);
 
 	private static final AppConstants APP_CONSTANTS = AppConstants.getInstance();
-	private File adapterFlowDir = new File(APP_CONSTANTS.getResolvedProperty("flow.adapter.dir"));
-	private File configFlowDir = new File(APP_CONSTANTS.getResolvedProperty("flow.config.dir"));
+	private File adapterFlowDir = new File(APP_CONSTANTS.getProperty("flow.adapter.dir"));
+	private File configFlowDir = new File(APP_CONSTANTS.getProperty("flow.config.dir"));
 	private @Setter ApplicationContext applicationContext;
 
 	private IFlowGenerator flowGenerator;
